Create Dockerfile and restructure directory
