--- elasticsearch_testing/elasticsearch-data/elasticsearch-response.js
@@ -1,10 +1,0 @@
-export const expectedResponse = 
-{
-    "total": 1,
-    "objects": [
-            {"_id":"5af9bb96083b8d0dae9e3c2a","name":"Six SCADA Modules","date":"1526486683452","length":"unit","levels":["undergraduate","community college"],"goals":[{"text":"<p>This project consists of six total modules with corresponding lab components, each divided based upon content and complexity. Modules are designed to use virtual SCADA testbeds in Docker, a program which runs on both Windows and Linux.&nbsp; The curriculum also includes lecture materials, homework, exam problems, and lab exercises. This curriculum has both two year college and four year college application.</p>\n"}],"published":true,"contributors":[],"collection":"nccp","status":"released","lock":null,"description":"<p>This project consists of six total modules with corresponding lab components, each divided based upon content and complexity. Modules are designed to use virtual SCADA testbeds in Docker, a program which runs on both Windows and Linux.&nbsp; The curriculum also includes lecture materials, homework, exam problems, and lab exercises. This curriculum has both two year college and four year college application.</p>\n","revision":0,"author":{"_id":"5b707232fd23f24b812e46e6","username":"alodgher","name":"akhtar lodgher","email":"akhtar.lodgher@tamusa.edu","organization":"texas a&m university- san antonio","objects":["5b70a01a7617f00001ef9fba","5b7aebb7f19e9800010ff93c","5b7d7c4d28e1640001a0c5d3","5b7d7e0528e1640001a0c5d5","5b7d7fd128e1640001a0c5da","5b7d8bce28e1640001a0c5e0","5b7d8d3b28e1640001a0c5e4","5b7d8e1328e1640001a0c5e7","5b7d8ef928e1640001a0c5e9","5b7d937228e1640001a0c5eb","5b7da30328e1640001a0c5f6","5b7da3cc28e1640001a0c5fa","5b7da50e28e1640001a0c5fd","5b7da61128e1640001a0c601","5b7da71928e1640001a0c604","5b7da7e528e1640001a0c607","5b7da88728e1640001a0c60a","5b85866136085c0001a4fdd4","5b85877836085c0001a4fdd7","5b85b1d036085c0001a4fdda","5b85b2d436085c0001a4fdde","5b85b3d136085c0001a4fde1","5b85bc3e36085c0001a4fde5","5b85bd2836085c0001a4fde9","5b85bdd636085c0001a4fdec","5b85beac36085c0001a4fdef","5b85bf7f36085c0001a4fdf3","5b85c05536085c0001a4fdf6","5b85c14536085c0001a4fdf9","5b85c1f236085c0001a4fdfc","5b85c2a436085c0001a4fe00","5b85c36236085c0001a4fe03","5b85c44836085c0001a4fe07","5b85c52a36085c0001a4fe0b","5b85c5c836085c0001a4fe0f","5b85c67536085c0001a4fe13","5b896b4a749b5f51604b58d5","5b896c57749b5fed6c4b58d9","5b896d74749b5f31414b58dc","5b896e1f749b5f2d514b58e5","5b896ebe749b5f03244b58e9"],"emailVerified":true,"bio":"","createdAt":"1534095922588"},"Learning-outcomes":[{"_id":"5af9bb96083b8d0dae9e3c2b","source":"5af9bb96083b8d0dae9e3c2a","tag":0,"author":"tommy morris","name":"Six SCADA Modules","date":"1526486683452","outcome":"Demonstrate Knowledge of computer networking concepts and protocols, and network security methodologies.","bloom":"Remember and Understand","verb":"Demonstrate","text":"Knowledge of computer networking concepts and protocols, and network security methodologies.","assessments":[],"strategies":[],"mappings":["5a674e2d04aa5f2a5ce97a8a","5a674e2d04aa5f2a5ce97c3d","5a674e2d04aa5f2a5ce97adb","5a674e2d04aa5f2a5ce97c72"]}],"5af9bb96083b8d0dae9e3c2a":"5af9bb96083b8d0dae9e3c2a"},
-            {"_id":"5ade6a02d93cc80db97019d7","name":"Prototyping","date":"1538513714059","length":"module","levels":["undergraduate"],"goals":[{"text":"<p>A prototype is an early sample, model, or release of a product built to test a concept or process or to act as a thing to be replicated or learned from. In this module, students will learn basic approaches to prototyping by building prototypes for a new Hue Bridge that supports users in making secure decisions. Students must work as a group to build a prototype of the device interface, the software interface, and the software functionality</p>\n"}],"published":true,"contributors":[],"collection":"nccp","status":"released","lock":null,"description":"<p>A prototype is an early sample, model, or release of a product built to test a concept or process or to act as a thing to be replicated or learned from. In this module, students will learn basic approaches to prototyping by building prototypes for a new Hue Bridge that supports users in making secure decisions. Students must work as a group to build a prototype of the device interface, the software interface, and the software functionality</p>\n","revision":0,"author":{"_id":"5ade4334c073ee18f7c69eca","username":"filiposharevski","name":"filipo sharevski","email":"fsharevs@cdm.depaul.edu","organization":"depaul university","objects":["5ade446bf9e4650da8998192","5ade4d41f9e4650da899819c","5ade5163f9e4650da89981a1","5ade55d6f9e4650da89981a9","5ade5902d93cc80db97019c3","5ade5c47d93cc80db97019cb","5ade5f65d93cc80db97019d2","5ade6a02d93cc80db97019d7","5adf4c156bb55e0daf9c0d80","5adf50d76bb55e0daf9c0d87","5b919a2ab2f96b03f87c86a1","5b9819375283d15fcf5b0a7a","5c05982ca3789358d6e7a624"],"emailVerified":true,"bio":"","accessGroups":[]},"Learning-outcomes":[{"_id":"5ade6ad3d93cc80db97019d8","source":"5ade6a02d93cc80db97019d7","tag":0,"author":"filipo sharevski","name":"Prototyping","date":"1538513714059","outcome":"Use user-centered design principles: discoverability, affordance, signifiers, feedback, constraint ","bloom":"Apply and Analyze","verb":"Use","text":"user-centered design principles: discoverability, affordance, signifiers, feedback, constraint ","assessments":[],"strategies":[],"mappings":[]},{"_id":"5ade702fd93cc80db97019d9","source":"5ade6a02d93cc80db97019d7","tag":1,"author":"filipo sharevski","name":"Prototyping","date":"1538513714059","outcome":"Use the design concepts of alignment, balance, order, contrast, proximity & similarity, rhythm, and white space","bloom":"Apply and Analyze","verb":"Use","text":"the design concepts of alignment, balance, order, contrast, proximity & similarity, rhythm, and white space","assessments":[],"strategies":[],"mappings":[]},{"_id":"5adf46c56bb55e0daf9c0d7a","source":"5ade6a02d93cc80db97019d7","tag":2,"author":"filipo sharevski","name":"Prototyping","date":"1538513714059","outcome":"Analyze what is emphasized in existing information design material (e.g. instructions & labeling in user interfaces) ","bloom":"Apply and Analyze","verb":"Analyze","text":"what is emphasized in existing information design material (e.g. instructions & labeling in user interfaces) ","assessments":[],"strategies":[],"mappings":[]},{"_id":"5adf477e6bb55e0daf9c0d7b","source":"5ade6a02d93cc80db97019d7","tag":3,"author":"filipo sharevski","name":"Prototyping","date":"1538513714059","outcome":"Analyze a user interface based on affordance and discoverability","bloom":"Apply and Analyze","verb":"Analyze","text":"a user interface based on affordance and discoverability","assessments":[],"strategies":[],"mappings":[]},{"_id":"5adf48de6bb55e0daf9c0d7c","source":"5ade6a02d93cc80db97019d7","tag":4,"author":"filipo sharevski","name":"Prototyping","date":"1538513714059","outcome":"Use the design concepts of alignment, balance, order, contrast, proximity & similarity, rhythm, and white space ","bloom":"Apply and Analyze","verb":"Use","text":"the design concepts of alignment, balance, order, contrast, proximity & similarity, rhythm, and white space ","assessments":[],"strategies":[],"mappings":["5a674e2d04aa5f2a5ce98000"]},{"_id":"5adf48f16bb55e0daf9c0d7d","source":"5ade6a02d93cc80db97019d7","tag":5,"author":"filipo sharevski","name":"Prototyping","date":"1538513714059","outcome":"Identify usability issues with devices and systems using Jakob Nielsen's 10 Usability Heuristics for User Interface Design","bloom":"Remember and Understand","verb":"Identify","text":"usability issues with devices and systems using Jakob Nielsen's 10 Usability Heuristics for User Interface Design","assessments":[],"strategies":[],"mappings":["5a674e2d04aa5f2a5ce97c40"]},{"_id":"5adf49e66bb55e0daf9c0d7e","source":"5ade6a02d93cc80db97019d7","tag":6,"author":"filipo sharevski","name":"Prototyping","date":"1538513714059","outcome":"Analyze the security flaws of a given Internet of Things workflow (application logic), including those introduced by the technology and those that could be introduced by an end user","bloom":"Apply and Analyze","verb":"Analyze","text":"the security flaws of a given Internet of Things workflow (application logic), including those introduced by the technology and those that could be introduced by an end user","assessments":[],"strategies":[],"mappings":["5a674e2d04aa5f2a5ce97b9d"]},{"_id":"5adf4b506bb55e0daf9c0d7f","source":"5ade6a02d93cc80db97019d7","tag":7,"author":"filipo sharevski","name":"Prototyping","date":"1538513714059","outcome":"Design a secure Internet of Things workflow (application logic) that eliminates these flaws ","bloom":"Evaluate and Synthesize","verb":"Design","text":"a secure Internet of Things workflow (application logic) that eliminates these flaws ","assessments":[],"strategies":[],"mappings":["5a674e2d04aa5f2a5ce97fff","5a674e2d04aa5f2a5ce98019"]}],"5ade6a02d93cc80db97019d7":"5ade6a02d93cc80db97019d7"},
-            {"_id":"5add2dc9e6b9490dd9331a37","name":"Adversarial Thinking","date":"1538511756267","length":"module","levels":["undergraduate","graduate","post graduate","community college","training"],"goals":[{"text":"<p>This curriculum module provides a basic introduction to adversarial thinking, game theory, and behavioral game theory to help develop cybersecurity students’ abilities to anticipate the strategic actions of cyber adversaries, including where, when, and how they might attack, and their tactics for evading detection. The basic premise of the module is that human adversaries are what differentiates cybersecurity from other technical disciplines such as computer science, and, therefore, the concept of adversarial thinking is central to cybersecurity. The goal of the module is to produce enduring strategic-mindedness in students who may otherwise tend to equate cybersecurity with technology-based best practices. This is a stand-alone, self-contained module, with no knowledge prerequisites. It contains three lessons of approximately one hour each. &nbsp;The module can be incorporated into virtually any university-level course. This module has been experimentally validated and is the subject of two peer-reviewed journal articles (cited in the syllabus).</p>\n"}],"published":true,"contributors":[],"collection":"nccp","status":"released","lock":null,"description":"<p>This curriculum module provides a basic introduction to adversarial thinking, game theory, and behavioral game theory to help develop cybersecurity students’ abilities to anticipate the strategic actions of cyber adversaries, including where, when, and how they might attack, and their tactics for evading detection. The basic premise of the module is that human adversaries are what differentiates cybersecurity from other technical disciplines such as computer science, and, therefore, the concept of adversarial thinking is central to cybersecurity. The goal of the module is to produce enduring strategic-mindedness in students who may otherwise tend to equate cybersecurity with technology-based best practices. This is a stand-alone, self-contained module, with no knowledge prerequisites. It contains three lessons of approximately one hour each. &nbsp;The module can be incorporated into virtually any university-level course. This module has been experimentally validated and is the subject of two peer-reviewed journal articles (cited in the syllabus).</p>\n","revision":0,"author":{"_id":"5add2c74c073ee18f7c69eb9","username":"shamman","name":"seth hamman","email":"shamman@cedarville.edu","organization":"cedarville university","objects":["5add2dc9e6b9490dd9331a37","5b81a11f36eac00001ea397b"],"emailVerified":true,"bio":""},"Learning-outcomes":[{"_id":"5add2f1ee6b9490dd9331a38","source":"5add2dc9e6b9490dd9331a37","tag":0,"author":"seth hamman","name":"Adversarial Thinking","date":"1538511756267","outcome":"Analyze a strategic scenario from the strategic perspective of cyber adversaries.","bloom":"Apply and Analyze","verb":"Analyze","text":"a strategic scenario from the strategic perspective of cyber adversaries.","assessments":[],"strategies":[],"mappings":["5b9ffaebe053b3382b3e93e4"]},{"_id":"5af9e102ec50aa101c13e5e1","source":"5add2dc9e6b9490dd9331a37","tag":1,"author":"seth hamman","name":"Adversarial Thinking","date":"1538511756267","outcome":"Apply level-k reasoning to derive playing strategies in strategic contests.","bloom":"Apply and Analyze","verb":"Apply","text":"level-k reasoning to derive playing strategies in strategic contests.","assessments":[],"strategies":[],"mappings":["5b9ffaebe053b3382b3e93e4"]},{"_id":"5af9e12dec50aa101c13e5e2","source":"5add2dc9e6b9490dd9331a37","tag":2,"author":"seth hamman","name":"Adversarial Thinking","date":"1538511756267","outcome":"Analyze a strategic scenario from a game theoretical perspective.","bloom":"Apply and Analyze","verb":"Analyze","text":"a strategic scenario from a game theoretical perspective.","assessments":[],"strategies":[],"mappings":["5b9ffaebe053b3382b3e93e4"]},{"_id":"5af9e15fec50aa101c13e5e3","source":"5add2dc9e6b9490dd9331a37","tag":3,"author":"seth hamman","name":"Adversarial Thinking","date":"1538511756267","outcome":"Illustrate the three components of adversarial thinking for cybersecurity.","bloom":"Remember and Understand","verb":"Illustrate","text":"the three components of adversarial thinking for cybersecurity.","assessments":[],"strategies":[],"mappings":["5b9ffaebe053b3382b3e93e4"]}],"5add2dc9e6b9490dd9331a37":"5add2dc9e6b9490dd9331a37"},
-            {"_id":"5aea58d2be129c0ccafc0d0e","name":"NICE Challenge - Malicious Malware!","date":"1527223900960","length":"micromodule","levels":["undergraduate","graduate","post graduate","community college"],"goals":[{"text":"<p>The NICE Challenges are hands-on scenarios designed to bring the workforce experience to students before they enter the workforce and are free&nbsp;to U.S. based educational institutions.</p>\n\n<p>Each challenge&nbsp;starts with a scenario written in the form of a task to complete from a fictional superior. For this module, it is as follows…</p>\n\n<p>\"We've noticed that our Domain-Controller has had some unauthorized usage by one of our interns and is now exhibiting some odd behavior. We believe it to be some type of malware that is causing the odd behavior. We need you to find the malware and remove it in order to prevent further harm to our system.\"</p>\n\n<p><strong>Lab Environment:</strong> The NICE Challenge Webportal is the lab environment used to complete and review NICE Challenges. The Webportal is a web platform which allows its&nbsp;users to interact with virtual environments and complete challenges requiring only a&nbsp;web browser. Details on how to sign-up as an educator and get started&nbsp;are in the notes section.</p>\n"}],"published":true,"contributors":[],"collection":"nccp","status":"released","lock":null,"description":"<p>The NICE Challenges are hands-on scenarios designed to bring the workforce experience to students before they enter the workforce and are free&nbsp;to U.S. based educational institutions.</p>\n\n<p>Each challenge&nbsp;starts with a scenario written in the form of a task to complete from a fictional superior. For this module, it is as follows…</p>\n\n<p>\"We've noticed that our Domain-Controller has had some unauthorized usage by one of our interns and is now exhibiting some odd behavior. We believe it to be some type of malware that is causing the odd behavior. We need you to find the malware and remove it in order to prevent further harm to our system.\"</p>\n\n<p><strong>Lab Environment:</strong> The NICE Challenge Webportal is the lab environment used to complete and review NICE Challenges. The Webportal is a web platform which allows its&nbsp;users to interact with virtual environments and complete challenges requiring only a&nbsp;web browser. Details on how to sign-up as an educator and get started&nbsp;are in the notes section.</p>\n","revision":0,"author":{"_id":"5ade428bc073ee18f7c69ec4","username":"jashley","name":"james ashley","email":"jashley@nice-challenge.com","organization":"nice challenge project | california state university, san bernardino","objects":["5ae8ed48be129c0ccafc0cd1","5aea4499be129c0ccafc0ce5","5aea4cf0be129c0ccafc0cea","5aea5177be129c0ccafc0cee","5aea5234be129c0ccafc0cf1","5aea53c1be129c0ccafc0cf4","5aea551bbe129c0ccafc0cf8","5aea56afbe129c0ccafc0cfd","5aea57a7be129c0ccafc0d06","5aea5852be129c0ccafc0d0a","5aea58d2be129c0ccafc0d0e","5aeba4b0be129c0ccafc0d11","5aeba612be129c0ccafc0d15","5aeba76cbe129c0ccafc0d1b","5aeba8d4be129c0ccafc0d22"],"emailVerified":true,"bio":""},"Learning-outcomes":[{"_id":"5aea58d2be129c0ccafc0d0f","source":"5aea58d2be129c0ccafc0d0e","tag":0,"author":"james ashley","name":"NICE Challenge - Malicious Malware!","date":"1527223900960","outcome":"Delete malware.","bloom":"Evaluate and Synthesize","verb":"Delete","text":"malware.","assessments":[],"strategies":[],"mappings":[]},{"_id":"5aea58d2be129c0ccafc0d10","source":"5aea58d2be129c0ccafc0d0e","tag":1,"author":"james ashley","name":"NICE Challenge - Malicious Malware!","date":"1527223900960","outcome":"Identify malware on the infected system.","bloom":"Apply and Analyze","verb":"Identify","text":"malware on the infected system.","assessments":[],"strategies":[],"mappings":["5a674e2d04aa5f2a5ce97abe","5a674e2d04aa5f2a5ce97aa1","5a674e2d04aa5f2a5ce97ad6","5a674e2d04aa5f2a5ce97bf7","5a674e2d04aa5f2a5ce97ee7","5a674e2d04aa5f2a5ce97afc","5a674e2d04aa5f2a5ce97be3","5a674e2d04aa5f2a5ce97c6d","5a674e2d04aa5f2a5ce97cae","5a674e2d04aa5f2a5ce97ebe","5a674e2d04aa5f2a5ce97aa7","5a674e2d04aa5f2a5ce97ab1"]}],"5aea58d2be129c0ccafc0d0e":"5aea58d2be129c0ccafc0d0e"},
-    ]
-}
