+++ elasticsearch_testing/gateways/learning-object-service-gateway.js
