+++ elasticsearch_testing/package.json
@@ -14,6 +14,7 @@
     "@elastic/elasticsearch": "^6.8.0",
     "dotenv": "^8.0.0",
     "jest": "^24.5.0",
-    "jsonwebtoken": "^8.5.1"
+    "jsonwebtoken": "^8.5.1",
+    "request-promise": "^4.2.4"
   }
 }
