--- elasticsearch_testing/response.txt
@@ -1,1 +1,0 @@
-{"total":10,"objects":[]}
